refactor(harmonize): extract color-math helpers to lib/color.mjs
Move the pure hex<->HSL / luminance / chromaticity / hue-distance helpers into
skills/autospec-harmonize/scripts/lib/color.mjs and import them, bringing
design-generalize.mjs from 432 to 380 LOC (under the AUTOSPEC_MAX_FILE_LOC=400
guardrail). No behavior change; 8/8 generalize tests still green.
