+++ skills/autospec-harmonize/scripts/design-generalize.mjs
@@ -28,6 +28,8 @@ import path from 'node:path';
 import { execSync } from 'node:child_process';
 import { fileURLToPath } from 'node:url';
 
+import { hexToHsl, luminance, isChromatic, hueDist } from './lib/color.mjs';
+
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 // ---------------------------------------------------------------------------
@@ -42,60 +44,6 @@ function parseArgs(argv) {
     }
   }
   return result;
-}
-
-// ---------------------------------------------------------------------------
-// Hex ↔ HSL conversion
-// ---------------------------------------------------------------------------
-
-/** Parse "#rrggbb" → [r, g, b] in [0,255]. */
-function hexToRgb(hex) {
-  const h = hex.replace('#', '');
-  return [
-    parseInt(h.slice(0, 2), 16),
-    parseInt(h.slice(2, 4), 16),
-    parseInt(h.slice(4, 6), 16),
-  ];
-}
-
-/** [r,g,b] in [0,255] → [h,s,l] where h∈[0,360), s∈[0,1], l∈[0,1]. */
-function rgbToHsl(r, g, b) {
-  r /= 255; g /= 255; b /= 255;
-  const max = Math.max(r, g, b);
-  const min = Math.min(r, g, b);
-  const l = (max + min) / 2;
-  if (max === min) return [0, 0, l]; // achromatic
-  const d = max - min;
-  const s = l > 0.5 ? d / (2 - max - min) : d / (max + min);
-  let h;
-  switch (max) {
-    case r: h = ((g - b) / d + (g < b ? 6 : 0)) / 6; break;
-    case g: h = ((b - r) / d + 2) / 6; break;
-    default: h = ((r - g) / d + 4) / 6; break;
-  }
-  return [h * 360, s, l];
-}
-
-/** Return [h, s, l] for a "#rrggbb" hex string. */
-function hexToHsl(hex) {
-  return rgbToHsl(...hexToRgb(hex));
-}
-
-/** Luminance proxy: l component from HSL (0=dark, 1=light). */
-function luminance(hex) {
-  return hexToHsl(hex)[2];
-}
-
-/** Is this hex "chromatic" (saturation >= 0.15)? */
-function isChromatic(hex) {
-  const [, s] = hexToHsl(hex);
-  return s >= 0.15;
-}
-
-/** Circular hue distance in degrees. */
-function hueDist(a, b) {
-  const d = Math.abs(a - b) % 360;
-  return d > 180 ? 360 - d : d;
 }
 
 // ---------------------------------------------------------------------------
